// src/types/diagnosisDTO.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Hopital Management System Api
 * OpenAPI spec version: v1
 */
import type { AppointmentDTO } from './appointmentDTO';

export interface DiagnosisDTO {
    appointment?: AppointmentDTO;
    appointmentId?: number;
    /** @nullable */
    description?: string | null;
    diagnosisDate?: string;
    diagnosisId?: number;
    /** @nullable */
    notes?: string | null;
}